+++ site/dat/docs/Changelog.md
@@ -3,6 +3,7 @@
 ## 0.4.6 (next)
  - allow using Xvfb for Selenium executor
  - don't use lynx browser when openin report links
+ - add PBench executor type
 
 ## 0.4.5
  - load settings applied to Stepping Thread Groups